+++ Nos/Models/EventReference+CoreDataClass.swift
@@ -14,4 +14,15 @@ public class EventReference: NosManagedObject {
     var jsonRepresentation: [String] {
         ["e", eventId ?? "", recommendedRelayUrl ?? "", marker ?? ""]
     }
+    
+    convenience init(jsonTag: [String], context: NSManagedObjectContext) throws {
+        guard jsonTag[safe: 0] == "e",
+            let eventID = jsonTag[safe: 1] else {
+            throw EventError.invalidETag(jsonTag)
+        }
+        self.init(context: context)
+        referencedEvent = try Event.findOrCreateStubBy(id: eventID, context: context)
+        recommendedRelayUrl = jsonTag[safe: 2]
+        marker = jsonTag[safe: 3]
+    }
 }