+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Render user avatar on Profile screen
 - Added support for threads in reply views
 - Retry failed Event sends every 2 minutes (max 5 retries)
+- Add basic notifications tab
 
 ## [0.1 (5)] 2023-03-02